add scroll view to the player details dialog
