+++ app/src/main/java/agersant/polaris/features/player/PlayerViewModel.kt
@@ -1,0 +1,7 @@
+package agersant.polaris.features.player
+
+import androidx.lifecycle.ViewModel
+
+class PlayerViewModel : ViewModel() {
+    var detailsShowing = false
+}
