+++ app/src/main/res/drawable/artwork_background.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <stroke
         android:width="@dimen/line_width"
-        android:color="?colorLine" />
+        android:color="?attr/colorDialogDivider" />
     <solid android:color="?android:colorBackground" />
     <corners android:radius="@dimen/artwork_radius" />
 </shape>
