+++ app/src/main/res/layout/view_player_controls.xml
@@ -187,7 +187,7 @@
         android:id="@+id/details"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_marginEnd="24dp"
+        android:layout_marginEnd="16dp"
         android:src="@drawable/ic_info_outline_24"
         app:layout_constraintBottom_toBottomOf="@id/next"
         app:layout_constraintEnd_toEndOf="parent"
