+++ app/src/main/res/values/attrs.xml
@@ -13,5 +13,5 @@
     <attr name="colorHeaderBackground" format="color" />
     <attr name="colorHeaderTextPrimary" format="color" />
     <attr name="colorHeaderTextSecondary" format="color" />
-    <attr name="colorLine" format="color" />
+    <attr name="colorDialogDivider" format="color" />
 </resources>
